Add implicit AnimatedAgent
If there is no AnimatedAgent in the parent hierarchy for an Animated
element, a global one "bound" on body will be used.
